Rebind to port even if existing TIME_WAIT still exist (like if PumpkinLB gets restarted). This will fix the "Cannot bind to port" errors when PumpkinLB is restarted after a connection has been made.
This uses the socket opt SO_REUSEADDR and may only work on UNIX.

The issue is, when PumpkinLB shuts down, an existing connection on that port keeps a socket open
at the kernel level in the "TIME_WAIT" state, to reduce overhead in creating sockets. Since it's no longer
running, we don't care and we want to just bind to the port anyway.